mailer/serializers.py
from rest_framework import serializers


class MailSerializer(serializers.Serializer):
    email = serializers.EmailField()
    subject = serializers.CharField()
    message = serializers.CharField()
