requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
flower==1.0.0
humanize==3.13.1
kombu==5.2.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pyparsing==3.0.6
pytz==2021.3
redis==4.1.1
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
